+++ formio/controllers/main.py
@@ -21,46 +21,31 @@ _logger = logging.getLogger(__name__)
 
 class FormioBaseController(http.Controller):
 
-    def _exception_submit(self, exception, submission_data, debug_mode=False):
-        _logger.error(exception)
-        if debug_mode and request.env.user.has_group('formio.group_formio_admin'):
-            se = http.serialize_exception(exception)
-            se['debug'] = traceback.format_exc()
-            error_lines = [
-                html_escape(se['debug']),
-                '',
-                html_escape(se['message'])
-            ]
-            error = '<br/>'.join(error_lines)
-            error = error.replace('\n', '<br/>')
-            error = error.replace('\\n', '<br/>')
-            error = error.replace('\\\n', '<br/>')
-            error = error.replace('\\\\n', '<br/>')
-        else:
-            # TODO configure message in form builder !
-            error = _('Sorry, something went wrong with processing the form. Please contact our support.')
-        # TODO message_post, create activity
-        return error
+    def _exception_load(self, exception, debug=False):
+        message = _('Sorry, something went wrong while loading the form. Please contact us.')
+        traceback_exc = traceback.format_exc()
+        _logger.error(traceback_exc)
+        traceback_exc_html = self._traceback_exc_html(exception, traceback_exc, debug=debug)
+        return message, traceback_exc_html
 
-    def _exception_submission(self, exception, submission_data, debug_mode=False):
-        _logger.error(exception)
-        if debug_mode and request.env.user.has_group('formio.group_formio_admin'):
-            se = http.serialize_exception(exception)
-            se['debug'] = traceback.format_exc()
-            error_lines = [
-                html_escape(se['debug']),
-                '',
-                html_escape(se['message'])
-            ]
-            error = '<br/>'.join(error_lines)
-            error = error.replace('\n', '<br/>')
-            error = error.replace('\\n', '<br/>')
-            error = error.replace('\\\n', '<br/>')
-            error = error.replace('\\\\n', '<br/>')
-        else:
-            error = _('Sorry, something went wrong with processing the form. Please contact our support.')
-        # TODO message_post, create actitivy
-        return error
+    def _exception_submit(self, exception, submission_data, debug=False):
+        message = _('Sorry, something went wrong while processing the form. Please contact us.')
+        traceback_exc = traceback.format_exc()
+        _logger.error(traceback_exc)
+        traceback_exc_html = self._traceback_exc_html(exception, traceback_exc, debug=debug)
+        return message, traceback_exc_html
+
+    def _traceback_exc_html(self, exception, traceback_exc, debug=False):
+        debug = debug or request.session.debug
+        se = http.serialize_exception(exception)
+        se['debug'] = traceback_exc
+        traceback_html = [html_escape(se['debug'])]
+        traceback_html = '<br/>'.join(traceback_html)
+        traceback_html = traceback_html.replace('\n', '<br/>')
+        traceback_html = traceback_html.replace('\\n', '<br/>')
+        traceback_html = traceback_html.replace('\\\n', '<br/>')
+        traceback_html = traceback_html.replace('\\\\n', '<br/>')
+        return traceback_html
 
     @http.route(['/web/content/<int:id>/fonts/<string:name>'], type='http', auth="public")
     def send_fonts_file(self, id, name, **kwargs):